Allow deprecation toolkit to be attached to multiple gems
Before, when the minitest plugin was used, the `attach_to` method of
the `DeprecationSubscriber` class was called first only with the
default set of gems to subscribe, which was `[:rails]`.

If the app added more gems to the `Configuration.attach_to` array,
the `attach_subscriber` method was called again with the new gems,
but this time, the `already_attached_to?` would return `true` since
there was already a notifier in the subscriber.

We should check the subscribers instead and make sure we don't attach
again to the same namespace.